+++ auth/permission.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 	"strings"
 
-	pkgctx "github.com/longvhv/saas-shared-go/context"
+	pkgctx "github.com/vhvcorp/go-shared/context"
 )
 
 var (